backend/requirements.txt
# Personal Assistant Backend Dependencies

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
asyncpg==0.29.0
sqlalchemy==2.0.23

# Authentication & Security
pyjwt==2.8.0
bcrypt==4.1.1
python-jose[cryptography]==3.3.0

# AI & Language Models
openai==1.3.7
pydantic==2.5.0

# External Integrations
google-auth==2.25.2
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.2.0
google-api-python-client==2.110.0

# Utilities
python-dotenv==1.0.0
httpx==0.25.2
redis==5.0.1
celery==5.3.4

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
mypy==1.7.1
